requirements.txt
flask==2.3.3
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.2.1
pyjwt
flask-middleware-jwt
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==20.1.0
psycopg2-binary
flask_cors
aiohttp